Week 4 Assignment for "Getting and Cleaning Data" Course
Contains the scripts, codebook and README files for the Week 4
assignment for cleaning data.  "Data_tidy.R" reads data files into R and
cleans the data to produce a tidy dataset containing averages of
measurement variables from the dataset "UCI HAR Dataset".  README.md is
a markdown file contaiing an overview of the files in this repository.